+++ src-next/tuning/equal.h
@@ -5,6 +5,8 @@
 #ifndef __SCXT_EQUAL_H
 #define __SCXT_EQUAL_H
 
+#include <cstddef>
+
 namespace scxt::tuning
 {
 struct EqualTuningProvider
@@ -16,16 +18,14 @@ struct EqualTuningProvider
      */
     float note_to_pitch(float note);
 
-
   protected:
-    static constexpr int tuning_table_size = 512;
+    static constexpr size_t tuning_table_size = 512;
     float table_pitch alignas(16)[tuning_table_size];
     float table_pitch_inv alignas(16)[tuning_table_size];
     float table_note_omega alignas(16)[2][tuning_table_size];
     // 2^0 -> 2^+/-1/12th. See comment in note_to_pitch
     float table_two_to_the alignas(16)[1001];
     float table_two_to_the_minus alignas(16)[1001];
-
 };
 extern EqualTuningProvider equalTuning;
 } // namespace scxt::tuning
